fix(bugs): another batch of fixes & inconsistencies (#90)
* fix(App): resolved a crash when accessing OOB index

* doc(TypedArgument): hidden + required shows nothing on help & throws an error

this is just documented. users do whatever they want with this
information.

* doc(App): fix footer() documentation

disable_footer_wrap is a bool looked at when help() is called so
irrelevant in order

* fix(HelpFormatter): wrapping wont cut open annotation's parenthesis

* fix(help): group("OPTIONS") joins the default section

* perf(help): stop copying the examples into HelpFormatter

* fix(help): drop the blank line and padding an empty description left

* fix(build): silence the remaining -Wconversion warnings

* ci(header-check): compile every src/ file standalone

* chore(clap.hpp): regen'd main header file

+++ src/damerau_osa.cpp
@@ -22,8 +22,8 @@ inline int clap::detail::damerau_osa(const std::string& s1, const std::string& s
     int cost = 0;
 
     // init d
-    for (i = 0; i < s1.size() + 1; i++) d[i][0] = i;
-    for (j = 0; j < s2.size() + 1; j++) d[0][j] = j;
+    for (i = 0; i < s1.size() + 1; i++) d[i][0] = static_cast<int>(i);
+    for (j = 0; j < s2.size() + 1; j++) d[0][j] = static_cast<int>(j);
 
     for (i = 1; i < s1.size() + 1; i++) {
         for (j = 1; j < s2.size() + 1; j++) {
